requirements.txt
torch>=1.12.0
numpy>=1.21.0
pandas>=1.3.0
scipy>=1.7.0
scikit-learn>=0.24.0
matplotlib>=3.4.0
seaborn>=0.11.0
shap>=0.40.0
networkx>=2.6.0
tqdm>=4.62.0
pyyaml>=5.4.0
tensorboard>=2.7.0
jupyterlab>=3.2.0
pytest>=6.2.0
black>=21.9b0
flake8>=3.9.0
mypy>=0.910